Enh: move *daemons*link(s) code/classes to shinken.objects.
