+++ shinken/objects/config.py
@@ -1,5 +1,4 @@
 #!/usr/bin/python
-
 # -*- coding: utf-8 -*-
 
 # Copyright (C) 2009-2014:
@@ -41,6 +40,7 @@ import cPickle
 import tempfile
 from StringIO import StringIO
 from multiprocessing import Process, Manager
+import json
 
 from item import Item
 from timeperiod import Timeperiod, Timeperiods
@@ -69,20 +69,19 @@ from hostextinfo import HostExtInfo, HostsExtInfo
 from serviceextinfo import ServiceExtInfo, ServicesExtInfo
 from trigger import Triggers
 from pack import Packs
-
 from shinken.util import split_semicolon
-from shinken.arbiterlink import ArbiterLink, ArbiterLinks
-from shinken.schedulerlink import SchedulerLink, SchedulerLinks
-from shinken.reactionnerlink import ReactionnerLink, ReactionnerLinks
-from shinken.brokerlink import BrokerLink, BrokerLinks
-from shinken.receiverlink import ReceiverLink, ReceiverLinks
-from shinken.pollerlink import PollerLink, PollerLinks
+from shinken.objects.arbiterlink import ArbiterLink, ArbiterLinks
+from shinken.objects.schedulerlink import SchedulerLink, SchedulerLinks
+from shinken.objects.reactionnerlink import ReactionnerLink, ReactionnerLinks
+from shinken.objects.brokerlink import BrokerLink, BrokerLinks
+from shinken.objects.receiverlink import ReceiverLink, ReceiverLinks
+from shinken.objects.pollerlink import PollerLink, PollerLinks
 from shinken.graph import Graph
 from shinken.log import logger
 from shinken.property import UnusedProp, BoolProp, IntegerProp, CharProp, StringProp, LogLevelProp, ListProp, ToGuessProp
 from shinken.daemon import get_cur_user, get_cur_group
 from shinken.util import jsonify_r
-import json
+
 
 no_longer_used_txt = 'This parameter is not longer take from the main file, but must be defined in the status_dat broker module instead. But Shinken will create you one if there are no present and use this parameter in it, so no worry.'
 not_interresting_txt = 'We do not think such an option is interesting to manage.'
