+++ test/test_poller_addition.py
@@ -21,8 +21,14 @@
 #
 # This file is used to test reading and processing of config files
 #
-
-from shinken_test import *
+import time
+from shinken.external_command import ExternalCommand
+from shinken.objects.brokerlink import BrokerLink
+from shinken.objects.arbiterlink import ArbiterLink
+from shinken.objects.pollerlink import PollerLink
+from shinken.objects.reactionnerlink import ReactionnerLink
+from shinken.objects.schedulerlink import SchedulerLink
+from shinken_test import ShinkenTest, unittest
 
 
 class GoodArbiter(ArbiterLink):
